Move widgets to ui module
